Added a "sort" parameter to pqgram_profile..
..because on re-reading the paper, the PQGram profiles are
not actually sorted (nor are they unique-ified!) in the suggested
algorithm. Sorting doesn't actually affect the score for the sample
trees, provided both trees are sorted. So, I've kept it as a feature
for cases where the tree structures don't guarantee child iteration
order, so that consistent results can be obtained in these cases.